refactor(ViabilityOfPaidDownloadingTorrent): moved
